backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
click==8.1.7
colorama==0.4.6
fastapi==0.115.5
frozenlist==1.5.0
h11==0.14.0
idna==3.10
multidict==6.1.0
propcache==0.2.1
pydantic==2.10.2
pydantic_core==2.27.1
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
yarl==1.18.3
